CREATE middleware validation.js and add it for Sessions
