+++ controllers/sessionController.js
@@ -1,5 +1,5 @@
 import Session from "../models/Session.js";
-import { formateDate } from "../middlewares/utils.js";
+import { formateDate } from "../middlewares/validation.js";
 
 /**
  *
